# Conflicts:
#	Binaries/Win64/UnrealEditor-Project_LLL.dll
#	Source/Project_LLL/Private/Entity/Object/Interactive/LLL_GateObject.cpp
#	Source/Project_LLL/Private/Entity/Object/Interactive/LLL_RewardObject.cpp
#	Source/Project_LLL/Private/Game/LLL_GameInstance.cpp
#	Source/Project_LLL/Private/System/MapGimmick/LLL_MapGimmick.cpp
#	Source/Project_LLL/Private/System/Reward/LLL_RewardGimmick.cpp
#	Source/Project_LLL/Private/UI/Entity/Character/Player/LLL_PlayerRewardInfoWidget.cpp
#	Source/Project_LLL/Public/Constant/LLL_FilePath.h
#	Source/Project_LLL/Public/Entity/Object/Interactive/LLL_GateObject.h
#	Source/Project_LLL/Public/Game/LLL_GameInstance.h
#	Source/Project_LLL/Public/UI/Entity/Character/Player/LLL_PlayerRewardInfoWidget.h

+++ Source/Project_LLL/Public/DataAsset/LLL_GoldComponentDataAsset.h
@@ -16,5 +16,5 @@ class PROJECT_LLL_API ULLL_GoldComponentDataAsset : public UDataAsset
 	
 public:
 	UPROPERTY(EditDefaultsOnly, Category = "Widget", DisplayName = "UI 노출 시간")
-	uint8 WidgetHideWaitTime;
+	float WidgetHideWaitTime;
 };
